adk/requirements.txt
google-adk==0.0.18
google-cloud-aiplatform
google-cloud-storage
google-cloud-bigquery
google-cloud-firestore
openai
